requirements.txt
pandas
numpy
akshare
matplotlib
arch
statsmodels
tqdm
pyyaml  # For reading the new config file